issueNumber fix

+++ app/src/main/java/com/demirci/marvel/data/source/model/character_details/Result.kt
@@ -53,7 +53,7 @@ data class Result(
     val issn: String?,
     @SerializedName("issueNumber")
     @Expose
-    val issueNumber: Int?,
+    val issueNumber: Double?,
     @SerializedName("modified")
     @Expose
     val modified: String?,
